refactor: split stats helper utilities
